requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=8.3.0
requests>=2.26.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
prometheus_client>=0.17.0
fastapi-cache2>=0.2.0
redis>=4.5.0
fastapi-throttling>=0.1.0 